backend/requirements.txt
fastapi
uvicorn
openai
transformers
torch
python-multipart
python-docx
PyPDF2
python-magic
anthropic
python-jose
passlib
pydantic